requirements.txt
colorama==0.3.9
PyBluez==0.22
PyDispatcher==2.0.5
six==1.11.0
smbus2==0.2.0
